+++ CHANGELOG.md
@@ -6,8 +6,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- N/A
 
+## [1.0.0]
+### Added
+- Add `<HeaderRow>` which is split into `left`, `center` and `right` tiers.
+- Add `<List>` section that supports a title and a description block.
+- Add simple `<ListRow>` with a Flexbox body for row components.
+- Add `<ColumnView>` which holds a `header` above and a `footer` below its main body area.
+- Add z-index `z()` sass helper. (Migrate from iC-framework)
+- Add `escapable()` mixin, listening `Esc` key to trigger `onEscape` prop.
+- Add page overlay component, `<Overlay>`.
+- Add `<Popup>` component.
+- Add 3 new icons for menu page purpose.
+
+### Changed
+- Improved tests coverage. (especially mixins & utils)
+- Font weight of `<body>` is now set as 400 by default.
+- `<Button>` active/hover colors are slightly darken.
+- `<Button>` now takes a `primary` prop to make it bolder.
+- `<EditableTextLabel>` only gets `autofocus` when it's also `inEdit`.
 
 ## [0.13.1]
 ### Added